# Optimize XxHash3