Hey there!
This is a PR to optimize XxHash3 for large buffers (> 240 bytes) by trying to avoid loading/storing the accumulators on each 64 bytes but instead doing it after a whole batch.

```
BenchmarkDotNet=v0.13.2, OS=Windows 11 (10.0.22000.1098/21H2)
AMD Ryzen 9 5950X, 1 CPU, 32 logical and 16 physical cores
.NET SDK=7.0.100-rc.1.22431.12
  [Host]     : .NET 7.0.0 (7.0.22.42610), X64 RyuJIT AVX2
  DefaultJob : .NET 7.0.0 (7.0.22.42610), X64 RyuJIT AVX2


|        Method |          data |     Mean |    Error |   StdDev | Ratio |
|-------------- |-------------- |---------:|---------:|---------:|------:|
|    XXH3Native | Byte[1048576] | 27.81 us | 0.070 us | 0.062 us |  0.43 |
|          XXH3 | Byte[1048576] | 64.14 us | 0.102 us | 0.085 us |  1.00 |
| XXH3Optimized | Byte[1048576] | 31.89 us | 0.070 us | 0.059 us |  0.50 |
```

- `XXH3Native` is the C++ version 
- `XXXH3` is the current C# version
- `XXH3Optimized` is this PR C# version

So overall, this PR brings x2 performance improvement on large buffers and lower the performance gap with the native version (~10%).

Small note not covered by this PR: while checking with the native version, I noticed that the hash is different (for both the current and optimized version vs native). Will have to dig further why.