@EgorBo 2022-11-01
> I noticed that the ARM64 implementation is significantly slower than the x86 version

I bet it's because of this: https://github.com/dotnet/runtime/pull/76641#discussion_r987378273

`operator *` for `Vector128<ulong>` ends up in a slow fallback because it doesn't exist on arm cc @stephentoub