@stephentoub 2022-11-02
> I noticed that the hash is different (for both the current and optimized version vs native). Will have to dig further why.

What native implementation are you calling and what are you comparing it against?  I'd validated all of the test data against a locally built xxhash.dll calling XXH3_64bits_withSeed.