@stephentoub 2022-11-02
> > I noticed that the ARM64 implementation is significantly slower than the x86 version
> 
> I bet it's because of this: [#76641 (comment)](https://github.com/dotnet/runtime/pull/76641#discussion_r987378273)
> 
> `operator *` for `Vector128<ulong>` ends up in a slow fallback because it doesn't exist on arm cc @stephentoub Namely, this line: https://github.com/stephentoub/runtime/blob/c3056f711fe315e71dc6e90ccd4184674bb01a73/src/libraries/System.IO.Hashing/src/System/IO/Hashing/XxHash3.cs#L784

Yeah, you can see the comment I left in the code about it and my question about better ideas for dealing with it:
https://github.com/dotnet/runtime/pull/76641#discussion_r987342213