@xoofx 2022-11-02
> What native implementation are you calling and what are you comparing it against? I'd validated all of the test data against a locally built xxhash.dll calling XXH3_64bits_withSeed.

Yes, I validated against a locally built xxhash.
Now, I was just casting to a `uint` but generating the full hash gives me the following:

```
Native: 0xd36c0e13a3df139e
C#:     0x9e13dfa3130e6cd3
```

The C# bytes are read from the destination span, casting to a `Span<ulong>`. The native version is what is returned by the API.

So the bytes are actually in reverse order. Is it expected?