@stephentoub 2022-11-02
> So the bytes are actually in reverse order. Is it expected?

Yes.
https://github.com/dotnet/runtime/blob/9a138bf5cb4a9055e4e0370a3eb3883514555fcd/src/libraries/System.IO.Hashing/src/System/IO/Hashing/XxHash3.cs#L21
https://github.com/dotnet/runtime/issues/76279

FYI, @bartonjs. I'm not alone here ;-)