@xoofx 2022-11-02
> Tests are failing, will dig further into these errors tonight.

Commit b72972c should fix the tests failing on the scalar path.