@stephentoub 2022-11-02
> XXHash is following the native endian order

I don't see any APIs that are outputting bytes, only numbers.  Can you elaborate?