@xoofx 2022-11-02
> I don't see any APIs that are outputting bytes, only numbers. Can you elaborate?

Sorry, I meant that they output numbers (except for the 128 bits which are 2x64 bit numbers) that thus follow the native endianness. Unless there is something in the .NET HashAlgorithm that mandates that all of them should always emit byte[] in big endian form?