@xoofx 2022-11-02
> "For systems which require to store and/or display the result in binary or hexadecimal format, the canonical format is defined to reproduce the same value as the natural decimal format, hence follows big-endian convention (most significant byte first)."

Oh, right, I can see it mentioned in the [header file here](https://github.com/Cyan4973/xxHash/blob/2de0fd6d9b3fd8bbb35ebd52152ecd11fc1ebbfc/xxhash.h#L661-L680)

That makes sense then, thanks!