@stephentoub 2022-11-03
FWIW, I don't see 50% improvement.  On my machine the best I see is ~25%:

| Method |         Toolchain |    Size |         Mean | Ratio |
|------- |------------------ |-------- |-------------:|------:|
|   Hash | \main\corerun.exe |     240 |     20.05 ns |  1.00 |
|   Hash |   \pr\corerun.exe |     240 |     20.10 ns |  1.00 |
|        |                   |         |              |       |
|   Hash | \main\corerun.exe |     512 |     31.55 ns |  1.00 |
|   Hash |   \pr\corerun.exe |     512 |     27.77 ns |  0.88 |
|        |                   |         |              |       |
|   Hash | \main\corerun.exe |    1024 |     52.49 ns |  1.00 |
|   Hash |   \pr\corerun.exe |    1024 |     41.17 ns |  0.78 |
|        |                   |         |              |       |
|   Hash | \main\corerun.exe | 1048576 | 83,650.94 ns |  1.00 |
|   Hash |   \pr\corerun.exe | 1048576 | 61,600.73 ns |  0.74 |

Still, a nice bump.  I don't love the manual loop unrolling and associated code increase, but it's probably still worth it.

Thanks.