@xoofx 2022-11-03
> FWIW, I don't see 50% improvement. On my machine the best I see is ~25%:

Oh interesting. What's your CPU?