@xoofx 2022-11-04
> ```
> BenchmarkDotNet=v0.13.2, OS=Windows 11 (10.0.22621.674)
> Intel Core i7-8700 CPU 3.20GHz (Coffee Lake), 1 CPU, 12 logical and 6 physical cores
> ``` 

Thanks, I don't have a recent Intel CPU, but I will check on a older Kaby-Lake how it behaves.

In the meantime, I received a Windows Dev Kit 2023 with ARM64 and the results are:

```
BenchmarkDotNet=v0.13.2, OS=Windows 11 (10.0.22621.755)
Snapdragon Compute Platform, 1 CPU, 8 logical and 8 physical cores
.NET SDK=7.0.100-rc.2.22477.23
  [Host]     : .NET 7.0.0 (7.0.22.47203), Arm64 RyuJIT AdvSIMD
  DefaultJob : .NET 7.0.0 (7.0.22.47203), Arm64 RyuJIT AdvSIMD


|        Method |          data |      Mean |    Error |   StdDev | Ratio |
|-------------- |-------------- |----------:|---------:|---------:|------:|
|    XXH3Native | Byte[1048576] |  70.16 us | 0.092 us | 0.086 us |  0.25 |
|          XXH3 | Byte[1048576] | 284.62 us | 0.882 us | 0.782 us |  1.00 |
| XXH3Optimized | Byte[1048576] | 156.05 us | 0.089 us | 0.074 us |  0.55 |
```

Which gives a similar performance boost to x2 times, but it is still 2x slower than the native version.